move checkpoint creation method to LogFileStorageService
