+++ rundeckapp/grails-app/services/rundeck/services/ExecutionService.groovy
@@ -1011,7 +1011,7 @@ class ExecutionService implements ApplicationContextAware, StepExecutor, NodeSte
                     jobcontext,
                     extraParamsExposed
             )
-            def checkpoint = workflowService.createPeriodicCheckpoint(execution)
+            def checkpoint = logFileStorageService.createPeriodicCheckpoint(execution)
 
             def wfEventListener = new WorkflowEventLoggerListener(executionListener)
             def logOutFlusher = new LogFlusher()
